refactor: useListbox to use Control.Listbox and EventHandlers
