--- src/DropdownListbox.bs.js
@@ -1,168 +1,0 @@
-// Generated by BUCKLESCRIPT, PLEASE EDIT WITH CARE
-'use strict';
-
-var Curry = require("bs-platform/lib/js/curry.js");
-var React = require("react");
-var Caml_obj = require("bs-platform/lib/js/caml_obj.js");
-
-function useDropdownListbox(labelId, optionsOpt, param) {
-  var options = optionsOpt !== undefined ? optionsOpt : [];
-  console.log("hey 1");
-  var match = React.useState(function () {
-        return false;
-      });
-  var setMenuVisible = match[1];
-  var menuVisible = match[0];
-  console.log("hey 2");
-  var match$1 = React.useState(function () {
-        return -1;
-      });
-  var setHighlightedIndex = match$1[1];
-  var highlightedIndex = match$1[0];
-  console.log("hey 3");
-  var match$2 = React.useState(function () {
-        return -1;
-      });
-  var setSelectedIndex = match$2[1];
-  var selectedIndex = match$2[0];
-  var size = options.length;
-  var labelId$1 = "id";
-  var highlightNext = function (param) {
-    return Curry._1(setHighlightedIndex, (function (param) {
-                  if ((size - param | 0) === 1) {
-                    return 0;
-                  } else {
-                    return param + 1 | 0;
-                  }
-                }));
-  };
-  var highlightPrev = function (param) {
-    return Curry._1(setHighlightedIndex, (function (param) {
-                  if (param === 0) {
-                    return size - 1 | 0;
-                  } else {
-                    return param - 1 | 0;
-                  }
-                }));
-  };
-  var showMenu = function (param) {
-    return Curry._1(setMenuVisible, (function (param) {
-                  return true;
-                }));
-  };
-  var hideMenu = function (param) {
-    return Curry._1(setMenuVisible, (function (param) {
-                  return false;
-                }));
-  };
-  var selectHighlighted = function (param) {
-    return Curry._1(setSelectedIndex, (function (param) {
-                  return highlightedIndex;
-                }));
-  };
-  var highlightFirst = function (param) {
-    return Curry._1(setHighlightedIndex, (function (param) {
-                  return 0;
-                }));
-  };
-  var highlightLast = function (param) {
-    return Curry._1(setHighlightedIndex, (function (param) {
-                  return size - 1 | 0;
-                }));
-  };
-  var getLabelProps = function (param) {
-    return {
-            id: labelId$1
-          };
-  };
-  var getToggleProps = function (param) {
-    return {
-            ariaLabelledby: labelId$1,
-            tabIndex: 0,
-            onClick: (function (param) {
-                param.preventDefault();
-                return Curry._1(showMenu, undefined);
-              }),
-            onKeyDown: (function (param) {
-                param.preventDefault();
-                var key = param.key;
-                switch (key) {
-                  case "ArrowDown" :
-                      if (menuVisible) {
-                        return Curry._1(highlightNext, undefined);
-                      } else {
-                        return Curry._1(showMenu, undefined);
-                      }
-                  case "ArrowUp" :
-                      if (menuVisible) {
-                        return Curry._1(highlightPrev, undefined);
-                      } else {
-                        return Curry._1(showMenu, undefined);
-                      }
-                  case "End" :
-                      if (menuVisible) {
-                        return Curry._1(highlightLast, undefined);
-                      } else {
-                        return ;
-                      }
-                  case " " :
-                  case "Enter" :
-                      if (menuVisible) {
-                        Curry._1(selectHighlighted, undefined);
-                        return Curry._1(hideMenu, undefined);
-                      } else {
-                        return Curry._1(showMenu, undefined);
-                      }
-                  case "Escape" :
-                      if (menuVisible) {
-                        return Curry._1(hideMenu, undefined);
-                      } else {
-                        return ;
-                      }
-                  case "Home" :
-                      if (menuVisible) {
-                        return Curry._1(highlightFirst, undefined);
-                      } else {
-                        return ;
-                      }
-                  default:
-                    return ;
-                }
-              }),
-            onBlur: (function (param) {
-                param.preventDefault();
-                return Curry._1(hideMenu, undefined);
-              })
-          };
-  };
-  var getMenuProps = function (param) {
-    return {
-            ariaLabelledby: labelId$1
-          };
-  };
-  var getOptionProps = function (index) {
-    return {
-            role: "option",
-            ariaSelected: Caml_obj.caml_equal(index, selectedIndex),
-            onClick: (function (param) {
-                return Curry._1(setSelectedIndex, (function (param) {
-                              return index;
-                            }));
-              })
-          };
-  };
-  return {
-          selectedIndex: selectedIndex,
-          highlightedIndex: highlightedIndex,
-          menuVisible: menuVisible,
-          getToggleProps: getToggleProps,
-          getLabelProps: getLabelProps,
-          getMenuProps: getMenuProps,
-          getOptionProps: getOptionProps,
-          showMenu: showMenu,
-          hideMenu: hideMenu
-        };
-}
-
-exports.useDropdownListbox = useDropdownListbox;
-/* react Not a pure module */
