+++ src/Listbox.bs.js
@@ -1,9 +1,30 @@
 // Generated by BUCKLESCRIPT, PLEASE EDIT WITH CARE
 'use strict';
 
-var DropdownListbox = require("./DropdownListbox.bs.js");
+var EventHandlers = require("./EventHandlers.bs.js");
+var Controls_Listbox = require("./Controls/Controls_Listbox.bs.js");
 
-var useDropdownListbox = DropdownListbox.useDropdownListbox;
+function useListbox(options) {
+  var size = options.length;
+  var match = Controls_Listbox.useControls(size);
+  var selectIndex = match.selectIndex;
+  var selectedIndex = match.selectedIndex;
+  var highlightedIndex = match.highlightedIndex;
+  var getOptionProps = function (index) {
+    return {
+            ariaSelected: selectedIndex === highlightedIndex,
+            role: "option",
+            onClick: (function (param) {
+                return EventHandlers.onClick(index, selectIndex, param);
+              })
+          };
+  };
+  return {
+          highlightedIndex: highlightedIndex,
+          selectedIndex: selectedIndex,
+          getOptionProps: getOptionProps
+        };
+}
 
-exports.useDropdownListbox = useDropdownListbox;
-/* DropdownListbox Not a pure module */
+exports.useListbox = useListbox;
+/* Controls_Listbox Not a pure module */
