--- src/Listbox.resi
@@ -1,5 +1,0 @@
-let useDropdownListbox: (
-  ~labelId: string=?, 
-  ~options: array<'a>=?,
-  unit
-) => DropdownListbox.dropdownListbox
